refactor: extract shared authedFetch helper
Move the duplicated authedFetch function from auth.ts, campaigns.ts,
and notifications.ts into a shared client.ts module.
