+++ packages/client/src/api/campaigns.ts
@@ -9,18 +9,7 @@ import type {
   CreateCampaignRequest,
   UpdateCampaignRequest,
 } from '@mmf/shared'
-
-function authedFetch(path: string, init: RequestInit = {}): Promise<Response> {
-  const token = localStorage.getItem('token')
-  const headers: Record<string, string> = {
-    'Content-Type': 'application/json',
-    ...(init.headers as Record<string, string>),
-  }
-  if (token) {
-    headers['Authorization'] = `Bearer ${token}`
-  }
-  return fetch(path, { ...init, headers })
-}
+import { authedFetch } from './client'
 
 export type {
   CampaignSummary,
