+++ packages/client/src/api/client.ts
@@ -1,0 +1,11 @@
+export function authedFetch(path: string, init: RequestInit = {}): Promise<Response> {
+  const token = localStorage.getItem('token')
+  const headers: Record<string, string> = {
+    'Content-Type': 'application/json',
+    ...(init.headers as Record<string, string>),
+  }
+  if (token) {
+    headers['Authorization'] = `Bearer ${token}`
+  }
+  return fetch(path, { ...init, headers })
+}
